Added simple mangrove tree (and some fixes) (#1349)
* Working on Improvements

* removed broken lines I cannot fix..

* Fixed brewingstand

Hoppers filled and draw items from the wrong slots. This is a fix for that.

* Fixed sneaking while flying

A Minecraft updated prevented sneaking while flying. It just does not send the start and stop sneaking actions anymore. But it still sends the SNEAKING action.
By asking for this, you can now place blocks next to interactable blocks while sneaking mid air

* Added annotions

* Sorry, forgot two annotions

* Requested changes