Fixed that exclusive enchantments was not compatible with custom items. (#2203)
This fix is to address exclusive enchantments for instance "Fire Protection" that can be only added to armors that was not compatible with custom items.

Also added enchantable slot type MACE, temporary supporting "all" type of enchantments for custom items while we don't know what exactly slot the client accepts for type mace.